Replace static_cast<int|i64>(...size()) with ssize() commit_hash:889f1aedbb4575db063289770c7a414cdc8d8fd6

+++ yt/yt/ytlib/chunk_client/block.h
@@ -32,7 +32,7 @@ struct TBlock
 
     explicit operator bool() const;
 
-    size_t Size() const;
+    i64 Size() const;
 
     TError ValidateChecksum() const;
